requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
Flask==2.3.1
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
pymongo==3.11.0
python-dotenv==1.0.0
Werkzeug==2.3.2
zipp==3.15.0
